move priority pkg to enterprise
