+++ enterprise/internal/insights/background/queryrunner/worker_test.go
@@ -11,12 +11,12 @@ import (
 	"github.com/hexops/autogold"
 
 	"github.com/sourcegraph/sourcegraph/enterprise/internal/insights/compression"
+	priority2 "github.com/sourcegraph/sourcegraph/enterprise/internal/insights/priority"
 	"github.com/sourcegraph/sourcegraph/enterprise/internal/insights/store"
 	"github.com/sourcegraph/sourcegraph/internal/actor"
 	"github.com/sourcegraph/sourcegraph/internal/database"
 	"github.com/sourcegraph/sourcegraph/internal/database/basestore"
 	"github.com/sourcegraph/sourcegraph/internal/database/dbtest"
-	"github.com/sourcegraph/sourcegraph/internal/insights/priority"
 )
 
 // TestJobQueue tests that EnqueueJob and dequeueJob work mutually to transfer jobs to/from the
@@ -132,7 +132,7 @@ func TestQueryExecution_ToQueueJob(t *testing.T) {
 		exec.Revision = "asdf1234"
 		exec.SharedRecordings = append(exec.SharedRecordings, bTime.Add(time.Hour*24))
 
-		got := ToQueueJob(&exec, "series1", "sourcegraphquery1", priority.Cost(500), priority.Low)
+		got := ToQueueJob(&exec, "series1", "sourcegraphquery1", priority2.Cost(500), priority2.Low)
 		autogold.Equal(t, got, autogold.ExportedOnly())
 	})
 	t.Run("test to job without dependents", func(t *testing.T) {
@@ -140,7 +140,7 @@ func TestQueryExecution_ToQueueJob(t *testing.T) {
 		exec.RecordingTime = bTime
 		exec.Revision = "asdf1234"
 
-		got := ToQueueJob(&exec, "series1", "sourcegraphquery1", priority.Cost(500), priority.Low)
+		got := ToQueueJob(&exec, "series1", "sourcegraphquery1", priority2.Cost(500), priority2.Low)
 		autogold.Equal(t, got, autogold.ExportedOnly())
 	})
 }
